feature(refactor): refactor code base
refactor client components to render pagination
split test according to controllers

[Finishes #156254322]
