+++ client/actionTypes/index.js
@@ -12,3 +12,4 @@ export const GET_USER_EVENTS = 'GET_USER_EVENTS';
 export const DELETE_EVENT = 'DELETE_EVENT';
 export const EDIT_EVENT = 'EDIT_EVENT';
 export const EDIT_CENTER = 'EDIT_CENTER';
+export const LOG_OUT = 'LOG_OUT';
