+++ client/components/Utils/Validator.js
@@ -1,0 +1,27 @@
+import { isEmpty } from 'lodash';
+
+export const validateSignup = (userData) => {
+  const errors = {};
+  if (userData.username === undefined || userData.username.trim() === '') {
+    errors.username = "Username is required."
+  }
+  if (userData.email === undefined || userData.email.trim() === '') {
+    errors.email = 'Email is required'
+  }
+  if (userData.password === undefined || userData.password.trim() === '') {
+    errors.password = 'Password is required'
+  }
+  if (userData.firstname === undefined || userData.firstname.trim() === '') {
+    errors.firstname = 'First name is required'
+  }
+  if (userData.lastname === undefined || userData.lastname.trim() === '') {
+    errors.lastname = 'Last name is required'
+  }
+  if (userData.password.length < 6) {
+    errors.password = 'Password should be atleast 6 characters long'
+  }
+  return {
+    errors,
+    validInput: isEmpty(errors)
+  }
+}
