+++ client/components/common/Navbar.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
+import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import decodeToken from '../../decodeToken'
+import * as userActions from '../../actions/actionCreator';
 
 class Navbar extends React.Component {
   constructor(props) {
@@ -9,12 +11,13 @@ class Navbar extends React.Component {
   }   
   logOut() {
     console.log(decodeToken, 'decode')
-    localStorage.clear()
+    localStorage.removeItem('x-access-token')
   
   }
   
 
   render() {
+    const { isAuthenticated } = this.props;
     return (
       <div>
         <header className="header">
@@ -27,23 +30,25 @@ class Navbar extends React.Component {
                   <li className="nav-item">                 <Link to="/" className="nav-link">Home</Link> 
                   </li>
                 </ul>
-                <ul className="navbar-nav ml-auto nav-flex-icons">
-                  <li className="nav-item">
-                    <Link to="/signup" className="nav-link waves-effect waves-light">Sign Up</Link>
-                  </li>
-                  <li className="nav-item">
-                    <Link to="/signin" className="nav-link waves-effect waves-light">Sign In</Link>
-                  </li>            
-                </ul>
-                <ul className="navbar-nav ml-auto nav-flex-icons">
-                  <li className="nav-item dropdown">
-                    <a className="nav-link dropdown-toggle waves-effect waves-light" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i className="fa fa-user" /> <strong>Admin</strong>
-                    </a>
-                    <div className="dropdown-menu dropdown-menu-right dropdown-unique" aria-labelledby="navbarDropdownMenuLink">
-                      <Link onClick={this.logOut} className="dropdown-item waves-effect btn btn-size waves-light ml-1 text-center" to="/">Log out</Link>
-                    </div>
-                  </li>
-                </ul>
+                {
+                  isAuthenticated ? <ul className="navbar-nav ml-auto nav-flex-icons">
+                    <li className="nav-item dropdown">
+                      <Link className="nav-link dropdown-toggle waves-effect waves-light" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" to='/addcenter'><i className="fa fa-user"  /> <strong>Admin</strong>
+                      </Link>
+                      <div className="dropdown-menu dropdown-menu-right dropdown-unique" aria-labelledby="navbarDropdownMenuLink">
+                        <Link onClick={this.props.logOut} className="dropdown-item waves-effect btn btn-size waves-light ml-1 text-center" to="/">Log out</Link>
+                      </div>
+                    </li>
+                  </ul> : <ul className="navbar-nav ml-auto nav-flex-icons">
+                    <li className="nav-item">
+                      <Link to="/signup" className="nav-link waves-effect waves-light">Sign Up</Link>
+                    </li>
+                    <li className="nav-item">
+                      <Link to="/signin" className="nav-link waves-effect waves-light">Sign In</Link>
+                    </li>            
+                  </ul>
+                }
+                
                 {/*  */}
                
               </div>
@@ -55,4 +60,16 @@ class Navbar extends React.Component {
   }
 }
 
-export default Navbar;
+export const mapStateToProps = state => {
+  console.log(">>>>>>>>>>>>state>>>>>>>", state);
+  return {
+    isAuthenticated: state.userAccess.isAuthenticated
+  }
+}
+export const mapDispatchToProps = dispatch => {
+  return {
+    logOut: userData => dispatch(userActions.logout({}))
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Navbar);
