+++ public/css/style.scss
@@ -183,6 +183,8 @@ a {
 .slot {
   margin-top: -44px;
 }
-// body {
-//   text-align: center;
-// }
+
+.max {
+  max-height: 245px;
+  min-height: 245px;
+}
