+++ server/test/index.js
@@ -1,2 +1,4 @@
-import './test';
+import './user/user';
+import './center/center';
+import './event/event'
 import './models/user';
