+++ webpack.config.js
@@ -1,4 +1,5 @@
 const htmlWebpackPlugin = require('html-webpack-plugin');
+const Dotenv = require('dotenv-webpack');
 const webpack = require('webpack');
 const path = require('path');
 
@@ -27,6 +28,9 @@ module.exports = {
   },
   plugins: [
     new webpack.HotModuleReplacementPlugin(),
+    new Dotenv({
+      systemvars: true
+    }),
     new htmlWebpackPlugin({
       template: path.resolve(__dirname, './client/index.html')
     })
